RemoteImage: allow addition of remote images.
This function is a wrapper around adding an image from a URL. It will
get the resposne (if successful) and ad its reader to the PDF. This can
later on be used with the `Image` method to position the image on the
page.

See the test file for a usage example.